test: Move commands test to commands module
