+++ src/commands/revert.rs
@@ -50,3 +50,47 @@ pub fn git_revert(commit_id: &str) -> Result<(), git2::Error> {
 
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::commands;
+    use crate::tests::get_repo;
+    use serial_test::serial;
+    use std::fs;
+
+    #[test]
+    #[serial]
+    fn test_git_revert() {
+        let repo = get_repo();
+
+        let file_name = "revert.txt";
+
+        // 파일 작성
+        fs::write(&file_name, "비빔밥").expect("failed to write file");
+        commands::git_add(file_name).expect("failed to add file");
+        let commit_msg = "비빔밥 먹고싶다.";
+        commands::git_commit(commit_msg).expect("failed to commit message");
+
+        let content = fs::read_to_string(file_name).expect("failed to read file");
+        assert_eq!(content, "비빔밥", "파일 생성 및 변경 안됨");
+
+        // 파일 수정
+        fs::write(file_name, "국밥").expect("failed to write file");
+        commands::git_add(file_name).expect("failed to add file");
+        let commit_msg = "비빔밥 질렸다.";
+        commands::git_commit(commit_msg).expect("failed to commit message");
+
+        let content = fs::read_to_string(file_name).expect("failed to read file");
+        assert_eq!(content, "국밥", "파일 변경 안됨");
+
+        // HEAD 커밋 id 가져오기
+        let head_commit = repo.head().expect("failed to get HEAD");
+        let commit_oid = head_commit.target().expect("HEAD refers to non-HEAD");
+
+        // git revert
+        commands::git_revert(&commit_oid.to_string()).expect("failed to revert");
+
+        let content = fs::read_to_string(file_name).expect("failed to read file");
+        assert_eq!(content, "비빔밥", "파일 롤백 안됨");
+    }
+}
